requirements.txt
Django==1.4
Jinja2==2.6
PyYAML==3.10
Pygments==1.5
Sphinx==1.1.3
beautifulsoup4==4.0.3
django-tastypie==0.9.11
django-celery==2.5.4
docutils==0.8.1
feedparser==5.1.1
gunicorn==0.14.2
lxml==2.3.4
mimeparse==0.1.3
git+git://github.com/winhamwr/mixpanel-celery.git#egg=mixpanel-celery
psycopg2==2.4.5
python-dateutil==1.5
git+https://github.com/NickPresta/pyuoguelph
uuid==1.30
wsgiref==0.1.2
yolk==0.4.3
